feat: add version field to bom, change price to float, add migration for bom versions table

+++ migrations/2024-04-03-214802_create_components/up.sql
@@ -8,6 +8,6 @@ CREATE TABLE components (
     part_number VARCHAR NOT NULL,
     description TEXT,
     supplier VARCHAR NOT NULL,
-    price_value INTEGER NOT NULL,
+    price_value REAL NOT NULL,
     price_currency VARCHAR NOT NULL
 );
